refactor(compress): Rename compress module to middleware/compress
The compress module has been moved to a new location under the middleware directory.
