+++ internal/app/middleware/compress/compress.go
