+++ internal/app/middleware/compress/compress_test.go
